% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnrichmentTable.R
\name{getEnrichmentTable}
\alias{getEnrichmentTable}
\title{Convert uniprot download to table protein entry and go column}
\usage{
getEnrichmentTable(enrich, background, mapping, topNodes = 15,
  ontology = "BP")
}
\arguments{
\item{enrich}{list of proteins}

\item{background}{background proteins}

\item{mapping}{default getHumanData() or getMouseData()}

\item{topNodes}{nr top categories}

\item{ontology}{ontology to use i.e. BB or CC}
}
\description{
Convert uniprot download to table protein entry and go column
}
\examples{

library(topGOUniProt)
tmp <- uniprotTable2ProteinGOTable(mapping=getMouseData(), ontology="GO")
head(tmp)
pids <- unique(tmp[,"Entry"] )
res <- getEnrichmentTable(pids[1:100], pids[1:1000] , mapping = getMouseData(0), topNodes = 15, ontology = "BP")
res
tmp <- uniprotTable2ProteinGOTable(mapping=getHumanData(),ontology="GO")
head(tmp)
pids <- unique(tmp[,"Entry"] )
res <- getEnrichmentTable(pids[1:2000], pids ,mapping = getMouseData(0), topNodes = 15, ontology = "MF")
res
res <- getEnrichmentTable(pids[1:2000], pids , getMouseData(0), topNodes = 15, ontology = "CC")
res

}

