% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMouseData.R
\name{getMouseData}
\alias{getMouseData}
\title{gets the mouse data annotations downloaded from uniprot}
\usage{
getMouseData()
}
\description{
gets the mouse data annotations downloaded from uniprot
}
\examples{
library(topGOUniProt)
mapping <- getMouseData()
head(mapping)
}

