% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topGODataNew.R
\name{topGODataNew}
\alias{topGODataNew}
\title{create an topGodata object}
\usage{
topGODataNew(selectedSet, backgroundSet, ontology = "BP",
  description = "Simple session", nodeSize = 5, mapping = getMouseData())
}
\description{
create an topGodata object
}
\examples{

tmp <-uniprotTable2ProteinGOTable(ontology="CC",mapping=getMouseData())
head(tmp)

protID <- unique(tmp[1:10,1])

protIDbackground <- unique(tmp[1:200,1])
res <- topGODataNew(selectedSet = protID , backgroundSet = protIDbackground, ontology="BP")
res <- topGODataNew(selectedSet = protID , backgroundSet = protIDbackground, ontology="CC")

summary(res)

}

