% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annFUN.R
\name{annFUN.uniprot}
\alias{annFUN.uniprot}
\title{topGo compatible annotation function}
\usage{
annFUN.uniprot(whichOnto, feasibleGenes, mapping)
}
\arguments{
\item{whichOnto}{- which ontology to use i.e. "BP", "CC"}

\item{feasibleGenes}{feasible Genes}

\item{mapping}{id go term mappings}
}
\description{
topGo compatible annotation function
}
\examples{
tmp <-uniprotTable2ProteinGOTable(mapping=getMouseData(), ontology="CC")
dim(tmp)
head(tmp)

protID <- unique(tmp[1:10,1])
t1 <- annFUN.uniprot("BP",protID, getMouseData(0) )
length(t1)
t1<-annFUN.uniprot("CC",protID, getMouseData(0) )
length(t1)
names(t1)
}

