% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnrichmentTable.R
\name{getEnrichmentTable}
\alias{getEnrichmentTable}
\title{Convert uniprot download to table protein entry and go column}
\usage{
getEnrichmentTable(enrich, background, topNodes = 15, ontology = "BP")
}
\arguments{
\item{enrich}{list of proteins}

\item{background}{background proteins}

\item{topNodes}{nr top categories}

\item{ontology}{ontology to use i.e. BB or CC}
}
\description{
Convert uniprot download to table protein entry and go column
}
\examples{

library(topGOUniProt)
tmp <- uniprotTable2ProteinGOTable(ontology="GO",mapping=getMouseData())
head(tmp)
pids <- unique(tmp[,"Entry"] )
res <- getEnrichmentTable(pids[1:100], pids[1:1000] , topNodes = 15, ontology = "BP")

tmp <- uniprotTable2ProteinGOTable(ontology="GO",mapping=getHumanData())
head(tmp)
pids <- unique(tmp[,"Entry"] )
res <- getEnrichmentTable(pids[sample(length(pids),2000)], pids , topNodes = 15, ontology = "CC")
res <- getEnrichmentTable(pids[1:2000], pids , topNodes = 15, ontology = "CC")
res


}

