% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHumanData.R
\name{getHumanData}
\alias{getHumanData}
\title{gets the human data annotations downloaded from uniprot}
\usage{
getHumanData()
}
\description{
gets the human data annotations downloaded from uniprot
}
\examples{
library(topGOUniProt)
mapping <- getHumanData()
head(mapping)

}

