% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHumanData.R
\name{getMouseData}
\alias{getMouseData}
\title{gets the mouse data annotations downloaded from uniprot}
\usage{
getMouseData(PE = 1)
}
\arguments{
\item{PE}{one of 0 = all, 1 = "Evidence at protein level", 2= "Evidence at transcript level", 3= "Inferred from homology",4="Predicted" , 5="Uncertain"}
}
\description{
gets the mouse data annotations downloaded from uniprot
}
\examples{
library(topGOUniProt)
tmp <- getMouseData()
colnames(tmp)
unique(tmp$PE)
}

