% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprotTable2ProteinGOTable.R
\name{uniprotTable2ProteinGOTable}
\alias{uniprotTable2ProteinGOTable}
\title{Convert uniprot download to table protein entry and go column}
\usage{
uniprotTable2ProteinGOTable(ontology = "GO", mapping = getMouseData())
}
\description{
Convert uniprot download to table protein entry and go column
}
\examples{

tmp <-uniprotTable2ProteinGOTable(ontology="GO",mapping=getMouseData())
dim(tmp)
head(tmp)
tmp <-uniprotTable2ProteinGOTable(ontology="BP",mapping=getMouseData())
dim(tmp)
head(tmp)
tmp <-uniprotTable2ProteinGOTable(ontology="CC",mapping=getMouseData())
dim(tmp)
head(tmp)

}

