% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprotTable2ProteinGOTable.R
\name{uniprotTable2ProteinGOTable}
\alias{uniprotTable2ProteinGOTable}
\title{Convert uniprot download to table protein entry and go column}
\usage{
uniprotTable2ProteinGOTable(mapping, ontology = "GO")
}
\arguments{
\item{mapping}{mapping}

\item{ontology}{ontology}
}
\description{
Convert uniprot download to table protein entry and go column
}
\examples{
library(topGOUniProt)

tmp <- uniprotTable2ProteinGOTable(ontology="GO",mapping=getMouseData(0))
dim(tmp)

}

